code/blog/requirements.txt
Django==2.2.2
djangorestframework==3.9.4
gunicorn==19.9.0